chore(wire-webapp, wire-team-settings): Remove translation files (#85)
BREAKING CHANGE:
Since we moved the translations to the app repositories there is no need to keep the translation files in the config anymore.

This used to be a shortcut to replace certain strings for self-hosted environments until we have everything string based covered with environment variables. That's the case now.